move compute_samples from optimizer to theseus_layer
